pip_requirements.txt
absl-py==2.1.0
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
black==24.8.0
blinker==1.8.2
cachetools==5.5.2
ccimport==0.3.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
crcmod==1.7
cryptography==44.0.2
cumm-cu111==0.2.9
cycler==0.12.1
dash==3.0.3
decorator==5.2.1
descartes==1.1.0
docker-pycreds==0.4.0
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.14.0
fire==0.7.0
flake8==7.1.2
flash-attn==0.2.2
Flask==3.0.3
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.12.3
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.7
lark==1.2.2
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
-e git+https://github.com/hanchaa/MEFormer.git@f4898ddf2fb825a89c932d2999cb4d7c468dacb9#egg=meformer
mmcls==0.25.0
mmcv-full==1.6.0
mmdet==2.24.0
mmdet3d==1.0.0rc5
mmsegmentation==0.29.1
model-index==0.1.11
motmetrics==1.4.0
mypy-extensions==1.0.0
narwhals==1.31.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
ninja==1.11.1.3
numba==0.53.0
numpy==1.23.5
nuscenes-devkit==1.1.11
oauthlib==3.2.2
open3d==0.19.0
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pccm==0.3.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
pluggy==1.5.0
plyfile==1.0.3
portalocker==3.0.0
prettytable==3.11.0
prompt_toolkit==3.0.51
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.1.4
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
rpds-py==0.20.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.27.0
setproctitle==1.3.5
Shapely==1.8.5.post1
six==1.17.0
smmap==5.0.2
spconv-cu111==2.1.25
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.2.1
torch==1.10.1+cu111
torchaudio==0.10.1+cu111
torchvision==0.11.2+cu111
tqdm==4.65.2
traitlets==5.14.3
trimesh==2.35.39
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wandb==0.19.10
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xmltodict==0.14.2
yapf==0.40.1
zipp==3.20.2
